Generalize unearthVersioner > ExecutableFileVersioner
